+++ engine/midi_builder.py
@@ -49,8 +49,8 @@ def build_midi_file(global_midi_data, file_name, metadata, output_dir):
         
         midi_events = []
         for note in events:
-            start_ticks = int(note['beat_position'] * ticks_per_beat)
-            end_ticks = int((note['beat_position'] + note['duration']) * ticks_per_beat)
+            start_ticks = int((note['beat_position'] + note.get('micro_offset', 0.0)) * ticks_per_beat)
+            end_ticks = int((note['beat_position'] + note.get('micro_offset', 0.0) + note['duration']) * ticks_per_beat)
 
             pitch = int(note['pitch'])